Suppression doublement des routes

+++ public/catalogue/routes/web.php
@@ -14,27 +14,21 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('medias', 'App\Http\Controllers\MediasController@getAllMedias');
-Route::get('{userId}/medias', 'App\Http\Controllers\MediasController@getAllMediasPlaylists')->middleware('auth')->name('medias.show');
-Route::post('{userId}/playlist/{playlistId}', 'App\Http\Controllers\PlaylistController@addMediaToPlaylist')->middleware('auth')->name('playlist.addMedia');
+Route::post('{userId}/{playlistId}', 'App\Http\Controllers\PlaylistController@addMediaToPlaylist')->middleware('auth')->name('playlist.addMedia');
 
 Route::get('films', 'App\Http\Controllers\MediasController@getAllFilms');
-Route::get('{userId}/films', 'App\Http\Controllers\MediasController@getAllFilmsPlaylists')->middleware('auth')->name('films.show');
-
 Route::get('series', 'App\Http\Controllers\MediasController@getAllSeries');
-Route::get('{userId}/series', 'App\Http\Controllers\MediasController@getAllSeriesPlaylists')->middleware('auth')->name('series.show');
-
 Route::get('search', 'App\Http\Controllers\SearchController@search');
 
 Route::get('addMedia', 'App\Http\Controllers\MediasController@formCreateMedia')->middleware('auth');
-Route::post('addMedia', 'App\Http\Controllers\MediasController@createMedia');
+Route::post('addMedia', 'App\Http\Controllers\MediasController@createMedia')->middleware('auth');
 
-Route::get('modifyMedia/{mediaId}', 'App\Http\Controllers\MediasController@formModifyMedia');
-Route::put('modifyMedia/{mediaId}', 'App\Http\Controllers\MediasController@modifyMedia');
+Route::get('modifyMedia/{mediaId}', 'App\Http\Controllers\MediasController@formModifyMedia')->middleware('auth');
+Route::put('modifyMedia/{mediaId}', 'App\Http\Controllers\MediasController@modifyMedia')->middleware('auth');
 
-Route::delete('deleteMedia/{mediaId}', 'App\Http\Controllers\MediasController@deleteMedia');
+Route::delete('deleteMedia/{mediaId}', 'App\Http\Controllers\MediasController@deleteMedia')->middleware('auth');
 
-Route::get('medias/{id}', 'App\Http\Controllers\MediasController@getMediaDetails');
-Route::get('{userID}/medias/{id}', 'App\Http\Controllers\MediasController@getMediaDetailsPlaylists')->middleware('auth')->name('medias.details');
+Route::get('medias/{id}', 'App\Http\Controllers\MediasController@getMediaDetails')->name('media.details');
 
 Route::get('medias/{id}/comments', 'App\Http\Controllers\CommentsController@getMediaComments');
 Route::post('medias/{id}/comments', 'App\Http\Controllers\CommentsController@postComment')->middleware('auth');
